# signfluent_phone/lib/src/api/bpm-service-api-spec.json
{"swagger":"2.0","info":{"description":"DESCRIPTION","version":"VERSION","title":"TITLE","termsOfService":"http://terms-of-services.url","license":{"name":"LICENSE","url":"http://url-to-license.com"}},"host":"localhost:8000","basePath":"/","tags":[{"name":"authentication-controller","description":"Authentication Controller"},{"name":"certificate-authority-controller","description":"Certificate Authority Controller"},{"name":"device-controller","description":"Device Controller"},{"name":"poc-controller","description":"POC Controller"},{"name":"role-management-controller","description":"Role Management Controller"},{"name":"signing-process-controller","description":"Signing Process Controller"}],"paths":{"/api/auth/createUser":{"post":{"tags":["authentication-controller"],"summary":"Creates a new user","operationId":"createUser","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userCreateRequest","description":"userCreateRequest","required":true,"schema":{"$ref":"#/definitions/UserCreateRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserCreateResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/auth/login":{"post":{"tags":["authentication-controller"],"summary":"Authenticates user","operationId":"login","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"authenticationRequest","description":"authenticationRequest","required":true,"schema":{"$ref":"#/definitions/AuthenticationRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AuthenticationResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/auth/ping":{"get":{"tags":["authentication-controller"],"summary":"Verifies the connection","operationId":"ping","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/auth/refreshToken":{"post":{"tags":["authentication-controller"],"summary":"Refreshes authentication token","operationId":"refreshAuthToken","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/RefreshTokenRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AuthenticationResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/certificateAuthority/issueX509Certificate":{"post":{"tags":["certificate-authority-controller"],"summary":"Issues X509 Certificate and returns in PEM format","operationId":"issueX509Certificate","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"issueX509CertificateRequest","description":"issueX509CertificateRequest","required":true,"schema":{"$ref":"#/definitions/IssueX509CertificateRequest"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/device/register":{"post":{"tags":["device-controller"],"summary":"Registers a new device","operationId":"register","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/RegisterDeviceRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RegisterDeviceResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/device/updateFCMToken":{"post":{"tags":["device-controller"],"summary":"Updates FCM token for specified device","operationId":"updateFCMToken","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"updateFCMTokenRequest","description":"updateFCMTokenRequest","required":true,"schema":{"$ref":"#/definitions/UpdateFCMTokenRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UpdateFCMTokenResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/poc/retrieve/{taskID}":{"post":{"tags":["poc-controller"],"summary":"fetchDocument","operationId":"fetchDocumentUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"taskID","in":"path","description":"taskID","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/poc/start":{"get":{"tags":["poc-controller"],"summary":"startProcess","operationId":"startProcessUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/poc/upload/{taskID}/{docID}":{"post":{"tags":["poc-controller"],"summary":"uploadDocumentId","operationId":"uploadDocumentIdUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"docID","in":"path","description":"docID","required":true,"type":"string"},{"name":"taskID","in":"path","description":"taskID","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/poc/uploadDocument":{"get":{"tags":["poc-controller"],"summary":"uploadDocument","operationId":"uploadDocumentUsingGET","produces":["*/*"],"parameters":[{"in":"body","name":"document","description":"document","required":true,"schema":{"$ref":"#/definitions/Document"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/role/assign":{"post":{"tags":["role-management-controller"],"summary":"Assign a role to the user","operationId":"assign","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"roleManagementRequest","description":"roleManagementRequest","required":true,"schema":{"$ref":"#/definitions/RoleManagementRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaseResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/role/getAll":{"post":{"tags":["role-management-controller"],"summary":"Returns all available roles","operationId":"getAll","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/SignfluentRoleResponse"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/role/getForCurrentUser":{"post":{"tags":["role-management-controller"],"summary":"Returns all roles for currently authenticated user","operationId":"getAllForCurrentUser","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/SignfluentRoleResponse"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/role/getForUser":{"post":{"tags":["role-management-controller"],"summary":"Returns roles for specified user","operationId":"getAlForUser","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userRequest","description":"userRequest","required":true,"schema":{"$ref":"#/definitions/UserRequest"}}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/SignfluentRoleResponse"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/role/revoke":{"post":{"tags":["role-management-controller"],"summary":"Revoke a role from the user","operationId":"revoke","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"roleManagementRequest","description":"roleManagementRequest","required":true,"schema":{"$ref":"#/definitions/RoleManagementRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaseResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/signingProcess/getActiveApprovers":{"post":{"tags":["signing-process-controller"],"summary":"Get active approvers","operationId":"getActiveApprovers","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/User"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/signingProcess/getDocumentDetails":{"post":{"tags":["signing-process-controller"],"summary":"Returns document details for specified task","operationId":"getDocumentDetails","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"taskIdRequest","description":"taskIdRequest","required":true,"schema":{"$ref":"#/definitions/TaskIdRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DocumentWithContent"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/signingProcess/getDocumentsForInspection":{"post":{"tags":["signing-process-controller"],"summary":"Returns documents needed inspection","operationId":"getDocumentsForInspection","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/DocumentResponse"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/signingProcess/inspectDocument":{"post":{"tags":["signing-process-controller"],"summary":"Inspect document and assign approvers","operationId":"inspectDocument","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"inspectDocumentRequest","description":"inspectDocumentRequest","required":true,"schema":{"$ref":"#/definitions/InspectDocumentRequest"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/signingProcess/uploadDocument":{"post":{"tags":["signing-process-controller"],"summary":"Uploads a document","operationId":"uploadDocument","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"uploadDocumentRequest","description":"uploadDocumentRequest","required":true,"schema":{"$ref":"#/definitions/UploadDocumentRequest"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"AuthenticationRequest":{"type":"object","properties":{"password":{"type":"string"},"username":{"type":"string"}},"title":"AuthenticationRequest"},"AuthenticationResponse":{"type":"object","properties":{"expiresIn":{"type":"integer","format":"int64"},"refreshExpiresIn":{"type":"integer","format":"int64"},"refreshToken":{"type":"string"},"token":{"type":"string"},"userId":{"type":"string"}},"title":"AuthenticationResponse"},"BaseResponse":{"type":"object","properties":{"message":{"type":"string"},"successful":{"type":"boolean"}},"title":"BaseResponse"},"DeviceEntity":{"type":"object","properties":{"deviceHardwareId":{"type":"string"},"fcmtoken":{"type":"string"},"id":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"}},"title":"DeviceEntity"},"Document":{"type":"object","properties":{"description":{"type":"string"},"status":{"type":"string"},"uploadedAt":{"type":"string","format":"date-time"},"uploadedBy":{"$ref":"#/definitions/User"}},"title":"Document"},"DocumentResponse":{"type":"object","properties":{"document":{"$ref":"#/definitions/Document"},"taskId":{"type":"string"}},"title":"DocumentResponse"},"DocumentWithContent":{"type":"object","properties":{"content":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"uploadedAt":{"type":"string","format":"date-time"},"uploadedBy":{"$ref":"#/definitions/User"}},"title":"DocumentWithContent"},"InspectDocumentRequest":{"type":"object","properties":{"approvers":{"type":"array","items":{"$ref":"#/definitions/UserDTO"}},"delegatorId":{"type":"string"},"isValid":{"type":"boolean"},"taskId":{"type":"string"}},"title":"InspectDocumentRequest"},"IssueX509CertificateRequest":{"type":"object","properties":{"publicKeyPem":{"type":"string"},"userId":{"type":"string"}},"title":"IssueX509CertificateRequest"},"RefreshTokenRequest":{"type":"object","properties":{"refreshToken":{"type":"string"}},"title":"RefreshTokenRequest"},"RegisterDeviceRequest":{"type":"object","properties":{"deviceHardwareId":{"type":"string"},"userId":{"type":"string","format":"uuid"}},"title":"RegisterDeviceRequest"},"RegisterDeviceResponse":{"type":"object","properties":{"deviceEntity":{"$ref":"#/definitions/DeviceEntity"},"message":{"type":"string"},"successful":{"type":"boolean"}},"title":"RegisterDeviceResponse"},"RoleManagementRequest":{"type":"object","properties":{"role":{"type":"string"},"userId":{"type":"string"}},"title":"RoleManagementRequest"},"SignfluentRoleResponse":{"type":"object","properties":{"key":{"type":"string"},"name":{"type":"string"}},"title":"SignfluentRoleResponse"},"TaskIdRequest":{"type":"object","properties":{"taskId":{"type":"string"}},"title":"TaskIdRequest"},"UpdateFCMTokenRequest":{"type":"object","properties":{"deviceHardwareId":{"type":"string"},"fcmToken":{"type":"string"},"userId":{"type":"string","format":"uuid"}},"title":"UpdateFCMTokenRequest"},"UpdateFCMTokenResponse":{"type":"object","properties":{"deviceEntity":{"$ref":"#/definitions/DeviceEntity"},"message":{"type":"string"},"successful":{"type":"boolean"}},"title":"UpdateFCMTokenResponse"},"UploadDocumentRequest":{"type":"object","properties":{"description":{"type":"string"},"document":{"type":"string"}},"title":"UploadDocumentRequest"},"User":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"id":{"type":"string"},"lastName":{"type":"string"},"username":{"type":"string"}},"title":"User"},"UserCreateRequest":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"username":{"type":"string"}},"title":"UserCreateRequest"},"UserCreateResponse":{"type":"object","properties":{"userId":{"type":"string"}},"title":"UserCreateResponse"},"UserDTO":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"id":{"type":"string"},"lastName":{"type":"string"},"username":{"type":"string"}},"title":"UserDTO"},"UserRequest":{"type":"object","properties":{"userId":{"type":"string"}},"title":"UserRequest"}}}